refactoring imports in app comp. moving profiles to components directory
